Radius Cull of Collision Check Objects (OPTIMIZATION)
Made it so that the player checks if an item is within range before checking it for collisions, if it is out of range it is skipped.
Doing so broke player movement because they would move so fast , other items would go from outside the radius to inside the player in a single frame.
Had to break resetAcceleration() into X and Y components to fix player base movement ont he ground so that I could lower the movespeed and keep the palyer from falling through the floor.
Unfortunately this introduced a new bug where some player input seems to result in random+rapid movement, which could cause the same collision bug to occur, meaning for now the culling range has to be very wide to accomodate.